feat(contents): add movie and tv show base models
